backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
supabase
google-genai
pydantic[email]
